chore: clean up of trade related slices and selectors (#6035)
* chore: remove unused selectors in swapper slice

* chore: merge swapperApi into swappersApi

* chore: rename swappersApi to swapeprApi

* chore: rename swappersSlice to tradeInputSlice

* chore: rename trade input selectors to include the word input for clarity

* chore: cleaned up and improved naming of trade selectors

* feat: bring bakk paranoia

* fix: merge

---------
