+++ src/components/MultiHopTrade/components/TradeInput/TradeInput.tsx
@@ -57,14 +57,14 @@ import {
   selectSwappersApiTradeQuotes,
   selectTradeQuoteRequestErrors,
   selectTradeQuoteRequestFailed,
-} from 'state/apis/swappers/selectors'
+} from 'state/apis/swapper/selectors'
 import {
-  selectBuyAsset,
+  selectInputBuyAsset,
+  selectInputSellAmountCryptoPrecision,
+  selectInputSellAsset,
   selectManualReceiveAddressIsValidating,
-  selectSellAmountCryptoPrecision,
-  selectSellAsset,
 } from 'state/slices/selectors'
-import { swappers } from 'state/slices/swappersSlice/swappersSlice'
+import { tradeInput } from 'state/slices/tradeInputSlice/tradeInputSlice'
 import {
   selectActiveQuote,
   selectActiveQuoteErrors,
@@ -118,8 +118,8 @@ export const TradeInput = memo(() => {
   const isKeplr = useMemo(() => !!wallet && isKeplrHDWallet(wallet), [wallet])
   const buyAssetSearch = useModal('buyAssetSearch')
   const sellAssetSearch = useModal('sellAssetSearch')
-  const buyAsset = useAppSelector(selectBuyAsset)
-  const sellAsset = useAppSelector(selectSellAsset)
+  const buyAsset = useAppSelector(selectInputBuyAsset)
+  const sellAsset = useAppSelector(selectInputSellAsset)
   const percentOptions = useMemo(() => {
     if (!sellAsset?.assetId) return []
     if (!isToken(fromAssetId(sellAsset.assetId).assetReference)) return []
@@ -136,7 +136,7 @@ export const TradeInput = memo(() => {
   const buyAmountAfterFeesUserCurrency = useAppSelector(selectBuyAmountAfterFeesUserCurrency)
   const totalNetworkFeeFiatPrecision = useAppSelector(selectTotalNetworkFeeUserCurrencyPrecision)
   const manualReceiveAddressIsValidating = useAppSelector(selectManualReceiveAddressIsValidating)
-  const sellAmountCryptoPrecision = useAppSelector(selectSellAmountCryptoPrecision)
+  const sellAmountCryptoPrecision = useAppSelector(selectInputSellAmountCryptoPrecision)
   const slippageDecimal = useAppSelector(selectTradeSlippagePercentageDecimal)
   const activeQuoteErrors = useAppSelector(selectActiveQuoteErrors)
   const quoteRequestErrors = useAppSelector(selectTradeQuoteRequestErrors)
@@ -162,15 +162,15 @@ export const TradeInput = memo(() => {
   }, [activeQuoteErrors, quoteRequestErrors])
 
   const setBuyAsset = useCallback(
-    (asset: Asset) => dispatch(swappers.actions.setBuyAsset(asset)),
+    (asset: Asset) => dispatch(tradeInput.actions.setBuyAsset(asset)),
     [dispatch],
   )
   const setSellAsset = useCallback(
-    (asset: Asset) => dispatch(swappers.actions.setSellAsset(asset)),
+    (asset: Asset) => dispatch(tradeInput.actions.setSellAsset(asset)),
     [dispatch],
   )
   const handleSwitchAssets = useCallback(
-    () => dispatch(swappers.actions.switchAssets()),
+    () => dispatch(tradeInput.actions.switchAssets()),
     [dispatch],
   )
 
@@ -180,7 +180,7 @@ export const TradeInput = memo(() => {
     dispatch(tradeQuoteSlice.actions.resetConfirmedQuote())
     // clear the active quote index on mount to prevent stale data affecting the selectors
     dispatch(tradeQuoteSlice.actions.resetActiveQuoteIndex())
-    dispatch(swappers.actions.setSlippagePreferencePercentage(undefined))
+    dispatch(tradeInput.actions.setSlippagePreferencePercentage(undefined))
   }, [dispatch])
 
   const {
