+++ src/components/MultiHopTrade/components/TradeSuccess/TradeSuccess.tsx
@@ -15,8 +15,8 @@ import { SlideTransition } from 'components/SlideTransition'
 import { RawText } from 'components/Text'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import { useLocaleFormatter } from 'hooks/useLocaleFormatter/useLocaleFormatter'
-import { selectEquityTotalBalance } from 'state/slices/selectors'
-import { selectLastHop, selectLastHopSellAccountId } from 'state/slices/tradeQuoteSlice/selectors'
+import { selectEquityTotalBalance, selectLastHopSellAccountId } from 'state/slices/selectors'
+import { selectLastHop } from 'state/slices/tradeQuoteSlice/selectors'
 import { useAppSelector } from 'state/store'
 
 import { TwirlyToggle } from '../TwirlyToggle'
