+++ src/components/MultiHopTrade/hooks/useIsTradingActive.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react'
 import type { ThorEvmTradeQuote } from 'lib/swapper/swappers/ThorchainSwapper/getThorTradeQuote/getTradeQuote'
 import { TradeType } from 'lib/swapper/swappers/ThorchainSwapper/utils/longTailHelpers'
-import { getIsTradingActiveApi } from 'state/apis/swapper/getIsTradingActiveApi'
-import { selectBuyAsset, selectSellAsset } from 'state/slices/swappersSlice/selectors'
+import { swapperApi } from 'state/apis/swapper/swapperApi'
+import { selectInputBuyAsset, selectInputSellAsset } from 'state/slices/tradeInputSlice/selectors'
 import { selectActiveQuote, selectActiveSwapperName } from 'state/slices/tradeQuoteSlice/selectors'
 import { useAppDispatch, useAppSelector } from 'state/store'
 
@@ -15,10 +15,10 @@ export const useIsTradingActive = () => {
 
   const dispatch = useAppDispatch()
 
-  const buyAssetId = useAppSelector(selectBuyAsset).assetId
-  const sellAssetId = useAppSelector(selectSellAsset).assetId
+  const buyAssetId = useAppSelector(selectInputBuyAsset).assetId
+  const sellAssetId = useAppSelector(selectInputSellAsset).assetId
 
-  const { getIsTradingActive } = getIsTradingActiveApi.endpoints
+  const { getIsTradingActive } = swapperApi.endpoints
   const swapperName = useAppSelector(selectActiveSwapperName)
 
   useEffect(() => {
