+++ src/state/apis/swapper/helpers/getInputOutputRatioFromQuote.test.ts
@@ -6,14 +6,14 @@ import { mockStore } from 'test/mocks/store'
 import { describe, expect, it, vi } from 'vitest'
 import { ETH } from 'lib/swapper/swappers/utils/test-data/assets'
 import { assertUnreachable } from 'lib/utils'
-import { getInputOutputRatioFromQuote } from 'state/apis/swappers/helpers/getInputOutputRatioFromQuote'
+import { getInputOutputRatioFromQuote } from 'state/apis/swapper/helpers/getInputOutputRatioFromQuote'
 import {
   cowQuote,
   lifiQuote,
   oneInchQuote,
   thorQuote,
   zrxQuote,
-} from 'state/apis/swappers/helpers/testData'
+} from 'state/apis/swapper/helpers/testData'
 import type { ReduxState } from 'state/reducer'
 
 const usdcAssetId: AssetId = 'eip155:1/erc20:0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48'
