+++ src/state/apis/swapper/helpers/validateTradeQuote.ts
@@ -19,11 +19,11 @@ import {
   selectAssets,
   selectFeeAssetById,
   selectFirstHopSellAccountId,
+  selectInputSellAmountCryptoPrecision,
   selectPortfolioAccountIdByNumberByChainId,
-  selectSellAmountCryptoPrecision,
+  selectSecondHopSellAccountId,
 } from 'state/slices/selectors'
 import { getTotalProtocolFeeByAssetForStep } from 'state/slices/tradeQuoteSlice/helpers'
-import { selectSecondHopSellAccountId } from 'state/slices/tradeQuoteSlice/selectors'
 
 import type { ErrorWithMeta } from '../types'
 import { type TradeQuoteError, TradeQuoteValidationError, TradeQuoteWarning } from '../types'
@@ -106,7 +106,7 @@ export const validateTradeQuote = async (
   const secondHop = quote.steps[1]
   const lastHop = isMultiHopTrade ? secondHop : firstHop
   const walletSupportedChainIds = selectWalletSupportedChainIds(state)
-  const sellAmountCryptoPrecision = selectSellAmountCryptoPrecision(state)
+  const sellAmountCryptoPrecision = selectInputSellAmountCryptoPrecision(state)
   const sellAmountCryptoBaseUnit = firstHop.sellAmountIncludingProtocolFeesCryptoBaseUnit
   const buyAmountCryptoBaseUnit = lastHop.buyAmountBeforeFeesCryptoBaseUnit
